Consolidate PubLicKeyMultibase generation
Use a single package for the generation
and parsing of the publicKeyMultibase's

This is a little tongue in cheek and must
be done correctly. Also we only ever want
to allow or deny keytypes.

In practise we only want RSA, but these
things change. If we add ones that are
only for signing and actually verification.

But our main focus is for encryption, RSA
is the goto for now.