Simplify encryption and signature keying
There is not need to add so much configurability. KISS.

We now resemple wireguard in terms of simplicity. That
feels better. By having a MIME type for both envelope
and message, we can change this later.
The MIME type adds a little girth in transport, but this seems
perfectly acceptable.