Remove MIMEType from envelope
Since the envelope isn't and shouldn't be signed,
we can trust the mime type and as such, it just
opens an attack vector. The only way seems to be
to just try and decrypt the thing and see if it
works.

Since the encryption method isn't likely to change
very often, if ever - this is OK.